backend/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
openai==1.52.2
supabase==2.10.0
websockets==12.0
anyio==4.6.2.post1
typing_extensions==4.12.2
starlette==0.38.6
